(Mixin)优化封禁系统处理

+++ src/main/kotlin/org/pio/rsn/model/UesrData.kt
@@ -3,7 +3,7 @@ package org.pio.rsn.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Integration(val count: Int, val time: Long, val active: Boolean, val nanoid: String)
+data class Integration(val count: Long, val time: Long, val active: Boolean, val nanoid: String)
 
 @Serializable
 data class Banned(val active: Boolean, val time: Long, val reason: String, val operator: String, val nanoid: String)
